minor code progress
