+++ examples/blueprint-construct/index.ts
@@ -3,6 +3,7 @@ import * as ec2 from "aws-cdk-lib/aws-ec2";
 import { CapacityType, KubernetesVersion, NodegroupAmiType } from 'aws-cdk-lib/aws-eks';
 import { Construct } from "constructs";
 import * as blueprints from '../../lib';
+//import { FlaggerAddOn } from '../../lib';
 import * as team from '../teams';
 
 const burnhamManifestDir = './examples/teams/team-burnham/';
@@ -124,9 +125,9 @@ export default class BlueprintConstruct extends Construct {
         });
 
         blueprints.EksBlueprint.builder()
-            .addOns(...addOns)
-            .clusterProvider(clusterProvider)
-            .teams(...teams)
+            //.addOns(new FlaggerAddOn())//...addOns)
+            //.clusterProvider(clusterProvider)
+            .teams()//...teams)
             .enableControlPlaneLogTypes(blueprints.ControlPlaneLogType.API)
             .build(scope, blueprintID, props);
     }
